expense-tracker-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
alembic>=1.16.0
python-dotenv>=1.0.0
email-validator>=2.2.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
joblib>=1.3.0
